more example content

+++ puavo-image-utils-0.1/opt/puavo-image-utils/example/more-parts/write/install.sh
@@ -3,23 +3,22 @@ cd $(dirname $0)
 
 echo *****************************write***************************************
 
-wget -O write-tgz http://www.styluslabs.com/download/write-tgz
+#download and unpack
+if ! test -d Write;then
+     test -f write.tgz || wget -O write-tgz http://www.styluslabs.com/download/write-tgz
+     tar xzf write-tgz
+     rm write-tgz
+fi
 
+#install
 mkdir -p /opt/Write/
-
-if test -d Write; then rm -rf Write; fi
-
-tar xzf write-tgz
 cd Write
-
 cp Write /opt/Write/.
 cp *.ttf /opt/Write/.
 cp Write144x144.png /usr/share/pixmaps/.
 cp Write.desktop /usr/share/applications/.
 
 
-rm write-tgz
-rm -r Write
